add: block user in chat + fix: problem with websocket in friends

+++ services/backend/user/tools/user/friends/views.py
@@ -9,6 +9,11 @@ from profiles.views import JWTAuthentication
 from profiles.models import CustomUser
 from friends.models import FriendRequest, BlockRequest
 from friends.serializers import FriendRequestSerializer, FriendsSerializer, BlockRequestSerializer
+from django.core.exceptions import PermissionDenied
+
+import logging
+
+logger = logging.getLogger(__name__)
 
 # class SendFriendRequestView(APIView):
 # 	authentication_classes = [JWTAuthentication]
@@ -102,41 +107,35 @@ class FriendsView(APIView):
 class BlockView(APIView):
 	authentication_classes = [JWTAuthentication]
 	def get(self, request, username=None):
-		if username:
-			block = get_object_or_404(request.user.block, username=username)
-			serializer = BlockSerializer(block)
-			return Response({'success': True, 'block': serializer.data}, status=status.HTTP_200_OK)
+		user = request.user
+		other_user = get_object_or_404(CustomUser, username=username)
+		user_blocked_other = BlockRequest.objects.filter(blocked=user, is_blocked=other_user).exists()
+		other_user_blocked_user = BlockRequest.objects.filter(blocked=other_user, is_blocked=user).exists()
+
+		if user_blocked_other or other_user_blocked_user:
+			return Response({
+				'success': True,
+				'user_blocked_other': user_blocked_other,
+				'other_user_blocked_user': other_user_blocked_user
+			}, status=status.HTTP_200_OK)
 		else:
-			return Response({'error': 'User not found in block list'}, status=status.HTTP_404_NOT_FOUND)
+			return Response({'success': False}, status=status.HTTP_200_OK)
 
 	def post(self, request, username=None):
-		from_user = request.user
-		to_user = get_object_or_404(request.user.block, username=username)
-		# friend_username = request.data.get('friend_username')
+		blocked = request.user
+		is_blocked = get_object_or_404(CustomUser, username=username)
 		try:
-			# to_user = CustomUser.objects.get(username=friend_username)
-			block_request, created = BlockRequest.objects.get_or_create(
-				from_user=from_user, to_user=to_user)
-
-			if created:
-				return Response({'success': True}, status=status.HTTP_201_CREATED)
-			else:
-				return Response({'error': 'Block request failed'}, status=status.HTTP_400_BAD_REQUEST)
-		except CustomUser.DoesNotExist:
-			return Response({'error': 'User not found'}, status=status.HTTP_404_NOT_FOUND)
+			BlockRequest.objects.create(blocked=blocked, is_blocked=is_blocked)
+			return Response({'success': True}, status=status.HTTP_201_CREATED)
 		except Exception as e:
 			return Response({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
 
 	def delete(self, request, username=None):
-		from_user = request.user
-		to_user = get_object_or_404(request.user.block, username=username)
-		try:
-			# to_user = CustomUser.objects.get(username=friend_username)
-			block_request = BlockRequest.objects.get(
-				from_user=to_user, to_user=from_user)
+		user = request.user
+		to_unblock = get_object_or_404(CustomUser, username=username)
+		block_request = BlockRequest.objects.filter(blocked=user, is_blocked=to_unblock).first()
+		if block_request:
 			block_request.delete()
 			return Response({'success': True}, status=status.HTTP_200_OK)
-		except BlockRequest.DoesNotExist:
-			return Response({'error': 'Block user not found'}, status=status.HTTP_404_NOT_FOUND)
-		except Exception as e:
-			return Response({'error': str(e)}, status=status.HTTP_400_BAD_REQUEST)
+		else:
+			raise PermissionDenied('You are not authorized to unblock this user')
